requirements.txt
# Core
streamlit>=1.35.0
python-dotenv>=1.0.0

# Database & Auth
sqlalchemy>=2.0.0
bcrypt>=4.1.0

# AI & Agents
langchain>=0.2.0
langchain-community>=0.2.0
langchain-anthropic>=0.1.0
langgraph>=0.1.0
anthropic>=0.30.0

# RAG & Vector Store
chromadb>=0.5.0

# Tools
requests>=2.31.0
reportlab>=4.2.0

# Utilities
pydantic>=2.0.0